Started working on the DatabasePool
We'll claim a certain amount of database connections (you can set this
in the config) and just keep them running. Later on you can simply request
a connection from the DatabasePool and a unused one will be returned (which
you'll have to release again later on). If you didn't set the max amount of
database connections it'll default to 10 which should work out just fine in
pretty much all cases, as the max amount of connections of MySQL is like 100.
Later on I'll likely build in a idling system into this pool where it'll close
connections if they're unused for more than %d seconds/minutes.

Signed-off-by: Toon Schoenmakers <nighteyes1993@gmail.com>