Started working on the verification of http requests signed with OAuth
Currently checking if all the expected fields are present and if the
consumer_token is present in the database. Also added a simple test
python script which you can find under test/test_oauth.py which will do
a simple signed request. As user tokens aren't implemented yet these are
just a couple of dummies.

Signed-off-by: Toon Schoenmakers <nighteyes1993@gmail.com>