Made the nonce check per user and some more improvements
Nonce now has a max length of 8, which is enforced by both the
database and our server (server will error you if it's longer).
Also keeping track of when the nonce was used, a nonce that is
in our cache for more than 5 minutes is no longer used in our
validation. It however isn't removed yet, we're still looking
into how we should do this. Also added a test which tests for
the too long nonce error.

Related to issue #1

Signed-off-by: Toon Schoenmakers <nighteyes1993@gmail.com>