Persona auth sort of working
There are a few issues. Mainly related to the callback url being called.
As I do the POST request to oauth/authorize from jquery now I have to
handle the 302 redirect from jquery as well. For some reason this seems
to happen twice (Chrome). But more important, the callback url doesn't
actually load in the web browser. It is just called in the background.
I'm everything but a javascript expert right here so I'm kind of clueless.

Signed-off-by: Toon Schoenmakers <nighteyes1993@gmail.com>