Added support for translations...
It has a couple of limitations for now though. It'll only 'resolve'
to the short names (so 'de' 'nl') and not the nl_NL etc. The default
(so no Accept-Language header or your language unavailable) is currently
our template page. Why you ask? Well there is no english translation yet :U

Signed-off-by: Toon Schoenmakers <nighteyes1993@gmail.com>